+++ src/javaTaskOne/guiMethods.java
@@ -2,17 +2,17 @@ package javaTaskOne;
 
 public class guiMethods {
 
-    public static int[] strArrayToIntArray(String[] a){
-        int[] b = new int[a.length];
+    public static short[] strArrayToIntArray(String[] a){
+        short[] b = new short[a.length];
         for (int i = 0; i < a.length; i++) {
-            b[i] = Integer.parseInt(a[i]);
+            b[i] = (short) Integer.parseInt(a[i]);
         }
 
         return b;
     }
 
-    public static int[][] arrayToMatrix (int[] intChars,int intN,int intM){
-        int[][] arr = new int[intN][intM];
+    public static short[][] arrayToMatrix (short[] intChars, int intN, int intM){
+        short[][] arr = new short[intN][intM];
         int count = 0;
 
         for (int i = 0; i < arr.length; i++) {
